refactor: RecordScreensRouter 파일 분리
